refactor: article processor into multiple files
